+++ packages/faucet/src/constants.ts
@@ -2,6 +2,7 @@ import { TokenConfiguration } from "./tokenmanager";
 import { parseBankTokens } from "./tokens";
 
 export const binaryName = "cosmwasm-faucet";
+export const memo: string | undefined = process.env.FAUCET_MEMO;
 export const concurrency: number = Number.parseInt(process.env.FAUCET_CONCURRENCY || "", 10) || 5;
 export const port: number = Number.parseInt(process.env.FAUCET_PORT || "", 10) || 8000;
 export const mnemonic: string | undefined = process.env.FAUCET_MNEMONIC;